v2.0.0 #9
Added actions:
+"SWS: Toolbar solo toggle" (Note, this replaces "SWS: Are any tracks soloed? [no-op, for toolbar]", please re-add to toolbar)
+"SWS: Toolbar arm toggle"
+"SWS: Set all sel tracks inputs to match first sel track"
Duplicate recording input check now ignores "input monitoring only" record mode.

+++ Autorender/Autorender.cpp
@@ -124,7 +124,7 @@ void CheckDirTree( const string &path, bool isFilename = false ){
 	bool skipRoot = false;
 #endif
 
-	unsigned int lastIndex = isFilename ? dirComps.size() - 2 : dirComps.size() - 1;
+	unsigned int lastIndex = (int)(isFilename ? dirComps.size() - 2 : dirComps.size() - 1);
 
 	for( unsigned int i = 0; i <= lastIndex; i++ ){
 		if( i > 0 ) buildPath += PATH_SLASH_CHAR;
